requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
schedule
python-dotenv
matplotlib
seaborn
jupyter
datetime
time